Get tickets, comments, audits, and add tests

+++ src/tl.employersupport.ecrm.poc.application/Interfaces/ITicketService.cs
@@ -5,6 +5,6 @@ namespace tl.employersupport.ecrm.poc.application.Interfaces
 {
     public interface ITicketService
     {
-        Task<Ticket> GetTicket(int ticketId);
+        Task<Ticket> GetTicket(long ticketId);
     }
 }
